requirements.txt
Django==4.1.1
djangorestframework==3.13.1
drf-yasg==1.21.3
psycopg2-binary==2.9.3
python-decouple==3.6
requests==2.28.1
stripe==4.1.0

